[code-infra] Convert @mui/styled-engine to TypeScript
