--- packages/mui-styled-engine/src/GlobalStyles/GlobalStyles.d.ts
@@ -1,11 +1,0 @@
-import * as React from 'react';
-import { Interpolation } from '@emotion/react';
-
-export interface GlobalStylesProps<Theme = {}> {
-  defaultTheme?: object | undefined;
-  styles: Interpolation<Theme>;
-}
-
-export default function GlobalStyles<Theme = {}>(
-  props: GlobalStylesProps<Theme>,
-): React.JSX.Element;
