+++ packages/mui-styled-engine/src/GlobalStyles/GlobalStyles.tsx
@@ -1,23 +1,34 @@
 'use client';
+import type * as React from 'react';
 import PropTypes from 'prop-types';
-import { Global } from '@emotion/react';
+import { Global, type Interpolation } from '@emotion/react';
 
-function isEmpty(obj) {
+export interface GlobalStylesProps<Theme = {}> {
+  defaultTheme?: object | undefined;
+  styles: Interpolation<Theme>;
+}
+
+function isEmpty(obj: object | null | undefined) {
   return obj === undefined || obj === null || Object.keys(obj).length === 0;
 }
 
-export default function GlobalStyles(props) {
+export default function GlobalStyles<Theme = {}>(
+  props: GlobalStylesProps<Theme>,
+): React.JSX.Element {
   const { styles, defaultTheme = {} } = props;
 
   const globalStyles =
     typeof styles === 'function'
-      ? (themeInput) => styles(isEmpty(themeInput) ? defaultTheme : themeInput)
+      ? (themeInput: Theme) =>
+          (styles as (theme: Theme) => Interpolation<Theme>)(
+            isEmpty(themeInput as object) ? (defaultTheme as Theme) : themeInput,
+          )
       : styles;
 
-  return <Global styles={globalStyles} />;
+  return <Global styles={globalStyles as any} />;
 }
 
-GlobalStyles.propTypes = {
+(GlobalStyles as any).propTypes /* remove-proptypes */ = {
   defaultTheme: PropTypes.object,
   styles: PropTypes.oneOfType([
     PropTypes.array,
