--- packages/mui-styled-engine/src/StyledEngineProvider/StyledEngineProvider.d.ts
@@ -1,9 +1,0 @@
-import * as React from 'react';
-
-export interface StyledEngineProviderProps {
-  children?: React.ReactNode;
-  enableCssLayer?: boolean | undefined;
-  injectFirst?: boolean | undefined;
-}
-
-export default function StyledEngineProvider(props: StyledEngineProviderProps): React.JSX.Element;
