+++ packages/mui-styled-engine/tsconfig.build.json
@@ -1,0 +1,17 @@
+// This config is for emitting declarations (.d.ts) only
+// Actual .ts source files are transpiled via babel
+{
+  "extends": "./tsconfig.json",
+  "compilerOptions": {
+    "composite": true,
+    "declaration": true,
+    "noEmit": false,
+    "emitDeclarationOnly": true,
+    "outDir": "build",
+    "rootDir": "./src",
+    "stripInternal": true,
+    "types": ["@mui/internal-code-infra/build-env", "react"]
+  },
+  "include": ["src/**/*.ts", "src/**/*.tsx"],
+  "exclude": ["src/**/*.test.ts*", "src/**/*.spec.ts*"]
+}
