+++ packages/mui-styled-engine/tsconfig.json
@@ -1,5 +1,9 @@
 {
   "extends": "../../tsconfig.json",
-
+  "compilerOptions": {
+    "allowJs": false,
+    "skipLibCheck": true,
+    "types": ["react", "vitest/globals", "@mui/internal-code-infra/build-env"]
+  },
   "include": ["src/**/*"]
 }
