+++ packages/mui-system/tsconfig.build.json
@@ -12,5 +12,8 @@
   },
   "include": ["src/**/*.ts*"],
   "exclude": ["src/**/*.spec.ts*", "src/**/*.test.ts*"],
-  "references": [{ "path": "../mui-utils/tsconfig.build.json" }]
+  "references": [
+    { "path": "../mui-styled-engine/tsconfig.build.json" },
+    { "path": "../mui-utils/tsconfig.build.json" }
+  ]
 }
